refactor : entity id 타입 변경 int -> Long  - 더 많은 수를 표현 하기 위함  - null값을 넣을 수 있음
refactor : createdDate, modifiedDate 리펙토링
BaseEntity.java - 추상 클래스에 createdDate, modifiedDate를 두고 상속해서 사용
DateTimeAttributeConverter.java - Entity와 DB 사이에서 createdDate, modifiedDate의 형변환 역할

+++ src/main/java/com/jig/blog/controller/BoardController.java
@@ -36,13 +36,14 @@ public class BoardController {
     }
 
     @GetMapping("/board/{id}")
-    public String getBoard(@PathVariable int id, Model model) {
+    public String getBoard(@PathVariable Long id, Model model) {
         model.addAttribute("board", boardService.getBoard(id));
+
         return "board/detail";
     }
 
     @GetMapping("/board/{id}/updateForm")
-    public String updateForm(@PathVariable int id, Model model) {
+    public String updateForm(@PathVariable Long id, Model model) {
         model.addAttribute("board", boardService.getBoard(id));
         return "board/updateForm";
     }
